+++ requirements.txt
@@ -1,7 +1,7 @@
-centrosome==1.0.2
+centrosome==1.0.3
 libtiff==0.4.0
 python-bioformats==1.0.5
-pyzmq==14.7.0
+pyzmq==13.1.0
 scikit-learn==0.16.1
 clint
 requests
