+++ setup.cfg
@@ -1,0 +1,3 @@
+[pytest]
+python_files = test_*
+testpaths = cellprofiler/tests cellprofiler/cpmath/tests cellprofiler/matlab/tests cellprofiler/modules/tests cellprofiler/gui/html/tests cellprofiler/utilities/tests imagej/tests
